[DevTools] Split MockRelayRenderer and renderRelayTree.
